Added LoginCheck and some JsonIgnores

+++ src/main/java/com/leonard/todo/repository/CityRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface CityRepository extends CrudRepository<City, Long> {
+public interface CityRepository extends CrudRepository<City, Integer> {
 
     List<City> findAll();
